Extract out Ui, Storage, Parser and TaskList classes
