+++ .gitignore
@@ -15,5 +15,6 @@ bin/
 
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
+/src/main/java/FakeDuke.java
 
 /src/main/java/*.class
