+++ src/main/java/DukeException.java
@@ -1,6 +1,10 @@
-public class NoStartException extends DukeException {
+public class DukeException extends Exception {
     private static String line = "--------------------------------------------------------------------";
-    public NoStartException() {
+
+    public DukeException() {
         super("(・´з`・) Uh oh... please add a start date\n" + line);
+    }
+    public DukeException(String message) {
+        super(message);
     }
 }
