+++ src/main/java/InvalidCommandException.java
@@ -1,4 +1,4 @@
-public class InvalidCommandException extends Exception {
+public class InvalidCommandException extends DukeException {
     private static String line = "--------------------------------------------------------------------";
     public InvalidCommandException() {
         super("Sorry, I don't understand what you mean ><\n" + line);
