+++ src/main/java/NoDescException.java
@@ -1,4 +1,4 @@
-public class NoDescException extends Exception {
+public class NoDescException extends DukeException {
     private static String line = "--------------------------------------------------------------------";
     public NoDescException() {
         super("(・´з`・) Uh oh... please add a description\n" + line);
