+++ src/main/java/NoEndException.java
@@ -1,4 +1,4 @@
-public class NoEndException extends Exception {
+public class NoEndException extends DukeException {
     private static String line = "--------------------------------------------------------------------";
     public NoEndException() {
         super("(・´з`・) Uh oh... please add an end date\n" + line);
